[Q20-15] - Finding patterns in the graph
